Added JavaScript libraries for initial compatibility and feature testing
This commit adds several JavaScript libraries to test their compatibility and features in the ePTS.Web project:

- bootstrap-datepicker
- bootstrap-icons
- datatables.net
- jstree
- select2
- select2-bootstrap-5-theme
- tom-select
- vanillajs-datepicker

Not all of these libraries will necessarily be kept long-term, as we are still evaluating which ones best fit our needs. The goal is to test them out and determine which to include in the final project.