Change AssessmentPerformanceLevel score to int
The AssessmentPerformanceLevel.Score property was originally defined as a double data type. However, this score simply captures the total number of learners that attained each performance level on an assessment.

Using an integer is sufficient for storing whole numbers for this purpose, and does not require decimal precision that double provides.

This commit changes the Score property on AssessmentPerformanceLevel from double to int to reduce memory usage and unnecessary precision.