Add multiple controllers, views and extensions
This commit creates the following new controllers and corresponding views:

- AssessmentPerformanceLevelsController
- AssessmentResultsController
- BaseController
- GradebookAssessmentsController
- GradebookEnrollmentsController
- GradebooksController
- HomeController
- OrganizationsController
- SchoolAcademicYearsController
- SchoolsController

This lays the foundation for the app by establishing the core controllers and views that will handle different areas of functionality. Additional actions and views can be added to these controllers as needed.

The PolicyTagHelper was also included to enable authorization policies.
The infopanel component was added for displaying headings and breadcrumbs.

Additionally, the following layout partial views were added:
- Layout
- LayoutBlank
- LayoutFooter
- LayoutNavBar
- LayoutSidebar
- LoginPartial
- ListScriptsPartial
- ListStylesPartial
- ValidationScriptsPartial

This lays the foundation for the app by establishing the core controllers, views, and layout components that will handle different areas of functionality. Additional actions and views can be added to these controllers as needed.

The HomeController and BaseController establish shared logic and layout. The other controllers and views will handle domain specific functionality for their respective models.

Added ApplicationLogs, Roles, AuditLogs, and Models