Implement Practice Environment Settings and Fix School Creation Bug
Practice Environment Configuration: Added PracticeEnvironment settings in appsettings.json to allow setup of a practice environment. This includes a toggle (IsActive) to activate or deactivate the practice mode. A message banner was added to the Layout when in practice mode.

The Serilog settings were removed from the application configuration.

Corrected an issue in the school create and edit views where the IsOrganizationUnit input field was incorrectly set to false. It is now correctly set to true, ensuring that new schools are created at the correct hierarchy level.